Moved Context Menu component
